requirements.txt
asgiref==3.2.10
certifi==2021.10.8
charset-normalizer==2.0.12
Django==3.1
django-crispy-forms==1.14.0
django-widget-tweaks==1.4.12
idna==3.3
Pillow==9.1.0
pytz==2022.1
requests==2.27.1
sqlparse==0.4.2
stripe==2.71.0
urllib3==1.26.9
